Module arti_client::config

Types and functions to configure a Tor client.

Some of these are re-exported from lower-level crates.

Modules

circ

Types for configuring how Tor circuits are built.

dir

Types for configuring how Tor accesses its directory information.

Macros

convert_helper_via_multi_line_list_builder

Implement TryFrom<MultilineListBuilder> and Into<MultilineListBuilder> for $Builder.

define_list_builder_accessors

Define accessor methods for a configuration item which is a list

define_list_builder_helper

Define a list builder struct for use with define_list_builder_accessors

impl_standard_builder

Defines standard impls for a struct with a Builder, incl Default

Structs

BridgeConfigBuilder

Configuration builder for a bridge - dummy type

BridgesConfig

Configuration for anti-censorship features: bridges and pluggable transports.

BridgesConfigBuilder

Builder for BridgesConfig.

CfgPath

A path in a configuration file: tilde expansion is performed, along with expansion of certain variables.

ChannelConfig

Channel configuration

ChannelConfigBuilder

Builder for ChannelConfig.

ClientAddrConfig

Configuration for client behavior relating to addresses.

ClientAddrConfigBuilder

Builder for ClientAddrConfig.

ConfigError

Wrapper for config::ConfigError with a more helpful error message.

MultilineListBuilderError

Error from trying to parse a MultilineListBuilder as a list of particular items

StorageConfig

Configuration for where information should be stored on disk.

StorageConfigBuilder

Builder for StorageConfig.

StreamTimeoutConfig

Configuration for client behavior relating to stream connection timeouts

StreamTimeoutConfigBuilder

Builder for StreamTimeoutConfig.

TorClientConfig

A configuration used to bootstrap a TorClient.

TorClientConfigBuilder

Builder for TorClientConfig.

Enums

BoolOrAuto

Boolean, but with additional "auto" option

CfgPathError

An error that has occurred while expanding a path.

ConfigBuildError

An error related to an option passed to Arti via a configuration builder.

ConfigurationSource

A configuration file or directory, for use by a ConfigurationSources

MultilineListBuilder

Configuration item specifiable as a list, or a single multi-line string

Reconfigure

Rules for reconfiguring a running Arti instance.

Constants

FS_PERMISSIONS_CHECKS_DISABLE_VAR

The environment variable we look at when deciding whether to disable FS permissions checking.

Functions

default_config_files

Return the filenames for the default user configuration files

fs_permissions_checks_disabled_via_envDeprecated

Return true if the environment has been set up to disable FS permissions checking.

Type Definitions

BridgeList

List of configured bridges, as found in the built configuration