Module arti_client::config::dir

Types for configuring how Tor accesses its directory information.

Structs

Authority

A single authority that signs a consensus directory.

AuthorityBuilder

Builder for Authority.

DirMgrConfig

Configuration type for network directory operations.

DirTolerance

Configuration for how much much to extend the official tolerances of our directory information.

DirToleranceBuilder

Builder for DirTolerance.

DownloadSchedule

Configuration for how many times to retry a download, with what frequency.

DownloadScheduleConfig

Configuration information for how exactly we download documents from the Tor directory caches.

DownloadScheduleConfigBuilder

Builder for DownloadScheduleConfig.

FallbackDir

A directory whose location ships with Tor (or arti), and which we can use for bootstrapping when we don’t know anything else about the network.

FallbackDirBuilder

Builder for FallbackDir.

NetworkConfig

Configuration information about the Tor network itself; used as part of Arti’s configuration.

NetworkConfigBuilder

Builder for NetworkConfig.