Enum arti_client::config::MultilineListBuilder

pub enum MultilineListBuilder<EB> {
    Unspecified,
    String(String),
    List(Vec<EB, Global>),
}

Configuration item specifiable as a list, or a single multi-line string

If a list is supplied, they are deserialized as builders. If a single string is supplied, it is split into lines, and #-comments and blank lines and whitespace are stripped, and then each line is parsed as a builder. (Eventually, the builders will be built.)

For use with sub_builder and define_list_builder_helper, with #[serde(try_from)] and #[serde(into)].

Example

use derive_builder::Builder;
use serde::{Deserialize, Serialize};
use tor_config::{ConfigBuildError, MultilineListBuilder};
use tor_config::convert_helper_via_multi_line_list_builder;
use tor_config::{define_list_builder_accessors, define_list_builder_helper};
use tor_config::impl_standard_builder;

#[derive(Debug, Clone, Builder, Eq, PartialEq)]
#[builder(build_fn(error = "ConfigBuildError"))]
#[builder(derive(Debug, Serialize, Deserialize))]
#[non_exhaustive]
pub struct LotteryConfig {
    /// What numbers should win the lottery?  Setting this is lottery fraud.
    #[builder(sub_builder, setter(custom))]
    #[builder_field_attr(serde(default))]
    winners: LotteryNumberList,
}
impl_standard_builder! { LotteryConfig }

/// List of lottery winners
//
// This type alias arranges that we can put `LotteryNumberList` in `LotteryConfig`
// and have derive_builder put a `LotteryNumberListBuilder` in `LotteryConfigBuilder`.
pub type LotteryNumberList = Vec<u16>;

define_list_builder_helper! {
    struct LotteryNumberListBuilder {
        numbers: [u16],
    }
    built: LotteryNumberList = numbers;
    default = generate_random();
    item_build: |number| Ok(*number);
    #[serde(try_from="MultilineListBuilder<u16>")]
    #[serde(into="MultilineListBuilder<u16>")]
}

convert_helper_via_multi_line_list_builder! {
    struct LotteryNumberListBuilder {
        numbers: [u16],
    }
}

define_list_builder_accessors! {
    struct LotteryConfigBuilder {
        pub winners: [u16],
    }
}

let lc: LotteryConfigBuilder = toml::from_str(r#"winners = [1,2,3]"#).unwrap();
let lc = lc.build().unwrap();
assert_eq!{ lc.winners, [1,2,3] }

let lc = r#"
winners = '''
  # Enny tells us this is the ticket they bought:

  4
  5
  6
'''
"#;
let lc: LotteryConfigBuilder = toml::from_str(lc).unwrap();
let lc = lc.build().unwrap();
assert_eq!{ lc.winners, [4,5,6] }

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Unspecified

Config key not present

String(String)

Config key was a string which is to be parsed line-by-line

List(Vec<EB, Global>)

Config key was a list of the individual entry builders