Trait arti_client::HasKind

pub trait HasKind {
    // Required method
    fn kind(&self) -> ErrorKind;
}

Errors that can be categorized as belonging to an ErrorKind

The most important implementation of this trait is arti_client::TorError; however, other internal errors throughout Arti also implement it.

Required Methods

fn kind(&self) -> ErrorKind

Return the kind of this error.